fixed last commit
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@5542 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/ANALYSIS/ID/PrecursorIonSelectionPreprocessing.C
@@ -369,7 +369,7 @@ namespace OpenMS
 								pt_map_.insert(make_pair(peptide_sequences[index2],detectabilities[index2]));
 							}
 						peptide_sequences.clear();
-						Int size = std::min(distance(seq_it,sequences_.end())-1,(int)max_peptides_per_run);
+						Int size = std::min((int)distance(seq_it,sequences_.end())-1,(int)max_peptides_per_run);
 						std::cout << "peptide_sequences.resize(size) "<<size<<std::endl;
  						peptide_sequences.resize(size);
 						index = 0;
